% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyBatteryOfMethods.R
\name{applyBatteryOfMethods}
\alias{applyBatteryOfMethods}
\title{Apply Battery of Methods}
\usage{
applyBatteryOfMethods(reports, n_drugs, n_events, tables = NULL,
  verbose = TRUE)
}
\arguments{
\item{reports}{The raw SR data set, a binary matrix}

\item{n_drugs, n_events}{The number of drugs and the number of events in the data set}

\item{tables}{A data frame with the 2x2 tables. 
Must contain the columns \code{drug_id}, \code{event_id}, 
\code{a}, \code{b}, \code{c} and \code{d}.
In case it is not given, it is derived from the reports.}

\item{verbose}{Verbosity (Default: \code{TRUE})}
}
\value{
The \code{tables} data frame with a column for each method
}
\description{
Applies a battery of methods to a simultaneous 
reporting data set
}
