% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchLASSO.R
\name{batchLASSO}
\alias{batchLASSO}
\title{Batch LASSO}
\usage{
batchLASSO(response, exposure, alpha = 1, verbose = TRUE)
}
\arguments{
\item{response}{A binary matrix where each column is a response variable}

\item{exposure}{A binary matrix where each column is an exposure}

\item{alpha}{The elastic net mixing parameter (Default: 1.0 - LASSO)}

\item{verbose}{Verbosity (Default: \code{TRUE})}
}
\value{
A data frame with three columns
        \item{response}{The response label (in case they are not given,
        they are simply numbered 1,2,3,...etc.)}
        \item{exposure}{The exposure label (in case they are not given,
        they are simply numbered 1,2,3,...etc.)}
        \item{highest_lambda}{The highest lambda for which the
        response-exposure pair where first added to the active set. The
        higher this value, the more "interesting" the pair}
}
\description{
This function can be used to rank the response-exposure pairs from most
to least "interesting". Simply using the resulting regression coefficients
would be unwise, since the size of the coefficients cannot be compared
from model to model. Instead, we use the tuning parameter, \eqn{\lambda}.
\cr\cr
Each response variable in the given \code{response} matrix is regressed
on all exposures in the matrix \code{exposure}. We determine for each
exposure what the highest value of \eqn{\lambda} for which that variable
is included for the first time in the regression model (i.e., its
regression coefficient is non-zero). These \eqn{\lambda}-values can be
compared across models.
}
\examples{
\dontrun{
n <- 100 # no. of observations
r <- 5 # no. of response variables
e <- 10 # no. of exposures

# random response and exposure matrices
response <- matrix(rbinom(r*n, 1, 0.5), n, r)
exposure <- matrix(rbinom(e*n, 1, 0.5), n, e)

batchLASSO(response, exposure)
}

}
