% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{Load Data}
\usage{
loadData(n_reports, n_drugs, n_events, alpha_drugs, beta_drugs,
  alpha_events, beta_events, n_correlated_drugs, rho_drugs,
  n_correlated_events, rho_events, n_correlated_pairs, rho,
  repetitions = c(1), type = "raw")
}
\arguments{
\item{n_reports}{Number of reports}

\item{n_drugs}{Number of drugs}

\item{n_events}{Number of adverse drug events}

\item{alpha_drugs}{Alpha parameter for the drug marginal probabilities}

\item{beta_drugs}{Beta parameter for the drug marginal probabilities}

\item{alpha_events}{Alpha parameter for the event marginal probabilities}

\item{beta_events}{Beta parameter for the event marginal probabilities}

\item{n_correlated_pairs}{Number of drug-event pairs that will exhibit a nonzero correlation}

\item{repetitions}{An array of repetitions one want to get}

\item{type}{Type of data. Takes the values \code{"raw"} for the raw data,
\code{"results"} for the table with the output of the methods in
the function \code{\link{applyBatteryOfMethods}} and \code{"auc"}
for the AUC filenames.}
}
\value{
data in a data frame. The column \code{repetition} represents the repetition
}
\description{
Returns the data from a number of repetitions.
}
