% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAUC.R
\name{plotAUC}
\alias{plotAUC}
\title{Boxplot of the AUCs}
\usage{
plotAUC(rho = 0.25, rho_drugs = 0, rho_events = 0, ylim = c(0.45, 0.9),
  title = "", type = "random", measure = "AUC")
}
\value{
A ROC plot
}
\description{
Returns a boxplot with the AUCs for each method.
}
