% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAUCBlock.R
\name{plotAUCBlock}
\alias{plotAUCBlock}
\title{Boxplot of the AUCs}
\usage{
plotAUCBlock(rho = 0.25, rho_drugs = 0, rho_events = 0, ylim = c(0.45,
  0.9), title = "")
}
\value{
A ROC plot
}
\description{
Returns a boxplot with the AUCs for each method.
}
