% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAUCDAG.R
\name{plotAUCDAG}
\alias{plotAUCDAG}
\title{Boxplot of the AUCs}
\usage{
plotAUCDAG(theta = 1.5, bystander_prob = 0.9, n_innocent_bystanders = 250,
  ylim = NULL, title = "", measure = "PRCAUC", font_size = 12,
  pattern = "aucdag+")
}
\value{
A ROC plot
}
\description{
Returns a boxplot with the AUCs for each method.
}
