% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAUCMethod.R
\name{plotAUCMethod}
\alias{plotAUCMethod}
\title{Boxplot of the AUCs}
\usage{
plotAUCMethod(method, ylim = NULL, title = "")
}
\value{
A ROC plot
}
\description{
Returns a boxplot with the AUCs for one particular method. 
It returns 3 boxplots with the results for all parameter settings
}
