% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAUCMethodBlock.R
\name{plotAUCMethodBlock}
\alias{plotAUCMethodBlock}
\title{Boxplot of the AUCs}
\usage{
plotAUCMethodBlock(method, ylim = NULL, title = "")
}
\value{
A ROC plot
}
\description{
Returns a boxplot with the AUCs for one particular method. 
It returns 3 boxplots with the results for all parameter settings
}
