% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelationMatrix.R
\name{plotCorrelationMatrix}
\alias{plotCorrelationMatrix}
\title{Plot the Correlation Matrix}
\usage{
plotCorrelationMatrix(corrmat, legend = TRUE, style = "blue")
}
\arguments{
\item{corrmat}{The correlation matrix.}
}
\value{
A heatmap of the correlation matrix
}
\description{
Creates a heatmap of the correlation matrix.
}
\details{
The code is based on
http://www.sthda.com/english/wiki/ggplot2-quick-correlation-matrix-heatmap-r-software-and-data-visualization
}
