% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROC.R
\name{plotROC}
\alias{plotROC}
\title{ROC plot}
\usage{
plotROC(n_reports = 1e+05, n_drugs = 500, n_events = 500,
  alpha_drugs = 1, beta_drugs = 5, alpha_events = 1, beta_events = 5,
  n_correlated_drugs = 50, rho_drugs = 0, n_correlated_events = 50,
  rho_events = 0, n_correlated_pairs = 100, rho = 0.25, repetition = 1,
  title = "ROC", method_set = "all")
}
\arguments{
\item{n_reports}{Number of reports}

\item{n_drugs}{Number of drugs}

\item{n_events}{Number of adverse drug events}

\item{alpha_drugs}{Alpha parameter for the drug marginal probabilities}

\item{beta_drugs}{Beta parameter for the drug marginal probabilities}

\item{alpha_events}{Alpha parameter for the event marginal probabilities}

\item{beta_events}{Beta parameter for the event marginal probabilities}

\item{n_correlated_pairs}{Number of drug-event pairs that will exhibit a nonzero correlation}
}
\value{
A ROC plot
}
\description{
Returns a ROC plot for a particular parameter setting
and a particular run (repetition)
}
