% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSettings.R
\name{plotSettings}
\alias{plotSettings}
\title{Plot for the Different Parameter Settings}
\usage{
plotSettings(method, rho = 1/4, ylim = NULL, method_name = method)
}
\value{
A plot
}
\description{
Returns a boxplot with the AUCs for one particular method.
}
