% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSettingsBlock.R
\name{plotSettingsBlock}
\alias{plotSettingsBlock}
\title{Plot for the Different Parameter Settings}
\usage{
plotSettingsBlock(method, datafile = "results_100000_100_100.rds",
  ylim = NULL, title = "")
}
\value{
A plot
}
\description{
Returns a boxplot with the AUCs for one particular method.
}
