% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSettingsDAG.R
\name{plotSettingsDAG}
\alias{plotSettingsDAG}
\title{Plot for the Different Parameter Settings}
\usage{
plotSettingsDAG(method, theta = 5, ylim = NULL, method_name = method,
  measure = "PRCAUC", return_table = FALSE)
}
\value{
A plot
}
\description{
Returns a boxplot with the AUCs for one particular method.
}
