% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnFilename.R
\name{returnFilename}
\alias{returnFilename}
\title{Simulation Data File Name for a DAG-based SR Data Set}
\usage{
returnFilename(n_reports, n_drugs, n_events, alpha_drugs, beta_drugs,
  alpha_events, beta_events, n_innocent_bystanders, theta_drugs,
  n_correlated_pairs, theta, repetitions, type = "raw")
}
\arguments{
\item{n_reports}{Number of reports}

\item{n_drugs}{Number of drugs}

\item{n_events}{Number of adverse drug events}

\item{alpha_drugs}{Alpha parameter for the drug marginal probabilities}

\item{beta_drugs}{Beta parameter for the drug marginal probabilities}

\item{alpha_events}{Alpha parameter for the event marginal probabilities}

\item{beta_events}{Beta parameter for the event marginal probabilities}

\item{theta_drugs}{The increase in odds-ratio when there is a connection between two drugs}

\item{n_correlated_pairs}{Number of drug-event pairs that will exhibit a nonzero correlation}

\item{theta}{Increase in odds-ratio when there is an edge going from a drug to an event}

\item{repetitions}{An array of repetitions one want to get}

\item{type}{Type of data. Takes the values \code{"raw"} for the raw data,
\code{"results"} for the table with the output of the methods in
the function \code{\link{applyBatteryOfMethods}} and \code{"auc"}
for the AUC filenames.}

\item{method}{The method used for generating the random DAG for the drugs}

\item{exp_degree}{Average degree between the drugs in the random DAG}
}
\value{
A vector of filenames
}
\description{
Returns the simulation data file name given a parameter setting
}
