% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{roc}
\alias{roc}
\title{Generate a ROC}
\usage{
roc(truth, output, decreasing = TRUE)
}
\arguments{
\item{truth}{Boolean vector with the true class}

\item{output}{The classifier's output}

\item{decreasing}{If \code{TRUE}, higher output is considered a stronger indication
for association than lower. For \eqn{p}-values, this should be 
\code{FALSE}. (Default: \code{TRUE})}
}
\value{
\item{TPR}{Vector with the true positive rate}
        \item{FPR}{Vector with the false positive rate} 
        \item{AUC}{Area under the curve - based on the values in \code{TPR} and \code{FPR}}
}
\description{
Returns the true positive rates (TPRs) and the false positive rates (FPRs)
that makes up the receiver operating curve (ROC). In case the classifier's 
output for a particular case is \code{NaN}, we assume that the classification
is \code{FALSE}, i.e., no association there. \cr\cr
This function is essentially a wrapper for the C++ function \code{\link{rorRcpp}}.
}
