% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rocRcpp}
\alias{rocRcpp}
\title{Generate a ROC}
\usage{
rocRcpp(truth, output, n_levels)
}
\arguments{
\item{truth}{A vector with the truth}

\item{output}{An order vector with the output of the method. The first elements
represent the items more likely to be TRUE.}

\item{n_levels}{Number of unique values in the output}
}
\value{
\item{TPR}{Vector with the true positive rate}
        \item{FPR}{Vector with the false positive rate} 
        \item{AUC}{Area under the curve - based on the values in \code{TPR} and \code{FPR}}
}
\description{
Generate a ROC
}
