% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateSmallestLambda}
\alias{updateSmallestLambda}
\title{Update the Tables with the Smallest Lambda}
\usage{
updateSmallestLambda(lowest_lambda, n_drugs, event_id, lambda, beta)
}
\arguments{
\item{lowest_lambda}{Vector with the lowest lambdas}

\item{n_drugs}{The number of drugs}

\item{event_id}{The ID of the event}

\item{lambda}{The shrinkage parameter that was used}

\item{beta}{The regression coefficients}
}
\value{
The \code{tables} with an updated column \code{smallest_lambda}
}
\description{
A function used by the \code{\link{batchLASSO}} function. 
It updates the column \code{smallest_lambda} in the data frame \code{tables}
with the regression estimates given in \code{beta}. The \code{beta}s are the 
result of regressing the drugs on event with id \code{event_id} using 
\code{lambda} as the shrinkage parameter.
}
